backend/requirements.txt
asgiref==3.6.0
Django==3.2.17
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.4.0